requirements.txt
anthropic>=0.40.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
